translator
It is a code that uses the Google Translate API to ensure the best translation results, and I also added the speaking voice feature through the gtts library. The translation of the code was through the googletrans library, and this and that the GUI was designed by the Pyqt5 library.